step count ios update ..
the refresh rate dropped from 20 to nearly 10, the lowest rate ive got so far.

thing i tried:
i created a custom step counter with a threshold multiplier, receiving data directly from the tracking api and sending it to the app directly, but it ended up to be slower than the api itself.